+++ ng/src/math.rs
@@ -1,2 +1,2 @@
-pub type Point = nalgebra::Point<f32, 3>;
-pub type Vector = nalgebra::SVector<f32, 3>;
+pub type Point = nalgebra::Point<f64, 3>;
+pub type Vector = nalgebra::SVector<f64, 3>;
